SCRUM-36 Added user to the is_authenticated backend response
SCRUM-36 The frontend only uses HTTPonly cookies. So for the frontend to check if a user is authenticated it sends a request to the backend on every URL change. The backend then verifies the cookies and returns the username and groups for the user. They are saved in state and passed along using the Auth context provider to child elements - like navbar, and dashboard - to verify user permissions.